requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# LSTM (forecast): CPU; tensorflow padrao em pip
tensorflow>=2.14.0

# MLflow
mlflow>=2.10.0
boto3>=1.28.0

# Visualization
matplotlib>=3.7.0
plotly>=5.24.0
seaborn>=0.12.0

# Utils
openpyxl>=3.1.0
pyarrow>=14.0.0
shap>=0.44.0
streamlit>=1.37.0
streamlit-option-menu>=0.4.0
